Split host strings in header file.
Since the hosts are stored in DNS wire format (with length-prefixed
labels), any labels that start with a digit cause an escape like
"\0071..." (that's length=7, first byte of the label = "1"). But this
causes a warning in Visual Studio and breaks the build.

This change splits the length bytes from the labels so that this warning
doesn't happen.